Move common functionality to BaseRepositoryTest
