+++ openlmis-stock-management-sample/src/androidTest/java/org/smartregister/stock/management/BaseRepositoryTest.java
@@ -1,9 +1,25 @@
 package org.smartregister.stock.management;
 
+import android.content.Context;
+import android.support.test.InstrumentationRegistry;
 import android.support.test.runner.AndroidJUnit4;
 
+import org.junit.BeforeClass;
 import org.junit.runner.RunWith;
+import org.smartregister.repository.Repository;
+import org.smartregister.stock.management.application.Application;
 
 @RunWith(AndroidJUnit4.class)
 public class BaseRepositoryTest {
+
+    protected static Context context;
+    protected static Repository mainRepository;
+
+    @BeforeClass
+    public static void bootStrap() {
+        context = InstrumentationRegistry.getTargetContext();
+        Application.setAppContext(context);
+        mainRepository = Application.getInstance().getRepository();
+    }
+
 }
