+++ openlmis-stock-management-sample/src/androidTest/java/org/smartregister/stock/management/TradeItemRepositoryTest.java
@@ -1,15 +1,10 @@
 package org.smartregister.stock.management;
 
-import android.content.Context;
-import android.support.test.InstrumentationRegistry;
 import android.util.Log;
 
 import org.junit.After;
 import org.junit.Before;
-import org.junit.BeforeClass;
 import org.junit.Test;
-import org.smartregister.repository.Repository;
-import org.smartregister.stock.management.application.Application;
 import org.smartregister.stock.management.domain.Gtin;
 import org.smartregister.stock.management.domain.TradeItem;
 import org.smartregister.stock.management.repository.TradeItemRepository;
@@ -22,16 +17,7 @@ import static org.smartregister.stock.management.util.Utils.DATABASE_NAME;
 
 public class TradeItemRepositoryTest extends BaseRepositoryTest {
 
-    private static Context context;
-    private static Repository mainRepository;
     private TradeItemRepository database;
-
-    @BeforeClass
-    public static void bootStrap() {
-        context = InstrumentationRegistry.getTargetContext();
-        Application.setAppContext(context);
-        mainRepository = Application.getInstance().getRepository();
-    }
 
     @Before
     public void setUp() {
